+++ demo/sat-demo/src/main/kotlin/ru/mipt/npm/sat/satServer.kt
@@ -10,8 +10,8 @@ import space.kscience.dataforge.meta.Null
 import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.dataforge.names.Name
 import space.kscience.visionforge.Colors
-import space.kscience.visionforge.html.Page
-import space.kscience.visionforge.html.plus
+import space.kscience.visionforge.html.VisionPage
+import space.kscience.visionforge.server.DataServeMode
 import space.kscience.visionforge.server.close
 import space.kscience.visionforge.server.openInBrowser
 import space.kscience.visionforge.server.serve
@@ -37,7 +37,8 @@ fun main() {
     }
 
     val server = satContext.visionManager.serve {
-        page(header = Page.threeJsHeader + Page.styleSheetHeader("css/styles.css")) {
+        dataMode = DataServeMode.UPDATE
+        page(VisionPage.threeJsHeader, VisionPage.styleSheetHeader("css/styles.css")) {
             div("flex-column") {
                 h1 { +"Satellite detector demo" }
                 vision { sat }