+++ ui/react/src/main/kotlin/space/kscience/visionforge/react/RangeValueChooser.kt
@@ -4,8 +4,8 @@ import kotlinx.css.pct
 import kotlinx.css.width
 import kotlinx.html.InputType
 import kotlinx.html.js.onChangeFunction
+import kotlinx.html.org.w3c.dom.events.Event
 import org.w3c.dom.HTMLInputElement
-import org.w3c.dom.events.Event
 import react.FC
 import react.dom.attrs
 import react.fc