+++ visionforge-markdown/src/commonMain/kotlin/space/kscience/visionforge/markup/MarkupPlugin.kt
@@ -1,0 +1,5 @@
+package space.kscience.visionforge.markup
+
+import space.kscience.visionforge.VisionPlugin
+
+public expect class MarkupPlugin: VisionPlugin